# [aten] Speed up vec256::loadu