Summary: Since we know we are loading fewer than 8 elements we can use vmaskmov, which is faster than memset.

Test Plan:
Microbenchmark on a skylake server shows an improvement from 17->10 ns.

Old and busted:
```
---------------------------------------------------------
Benchmark                  Time           CPU Iterations
---------------------------------------------------------
Loadu                     17 ns         17 ns   40529728
```

New hotness:
```
---------------------------------------------------------
Benchmark                  Time           CPU Iterations
---------------------------------------------------------
Loadu                     10 ns         10 ns   68226581
```

Differential Revision: D29699470