@bertmaher 2021-08-23
abandoning this, b/c there are possible memory safety issues with masked loads on some platforms and it's not worth trying to test this absolutely everywhere.