@colesbury 2021-08-25
What are the memory safety issues? On which platofrms?