@bertmaher 2021-08-25
Sorry @colesbury I should've put this finding on the the public PR.  The AMD manual leaves open the possibility that masked-out addresses can still generate page faults:

I usually just read the Intel docs for fault behavior, which is reassuring:
> Faults occur only due to mask-bit required memory accesses that caused the faults. Faults will not occur due to referencing any memory location if the corresponding mask bit for that memory location is 0. For example, no faults will be detected if the mask bits are all zero.

but the AMD docs crush my hopes:
> Exception and trap behavior for elements not selected for loading or storing from/to memory is implementation dependent. For instance, a given implementation may signal a data breakpoint or a page fault for doublewords that are zero-masked and not actually written.
So, damn.  Seems like this could cause a spurious page fault/SEGV if it's used unaligned at the boundary of an unmapped page 

I am sort of skeptical whether it would actually be a problem in practice but it also feels like a bit of a risk for a very small micro-optimization